Moved Vega-lite/Hanami to new file, modified line labels
Previously I assumed that all mapping lines would start from
same location, but I want to be able to have multiple mapping lines
one one plot, and different mapping lines on different plots.  So
now the init-x value that characterizes a mapping line is now
displayed as part of the label of the mapping line.  Prevously it
was part of the label for the curve, which didn't make sense (but
was succinct).