Swapped K and r params.
The idea is that if there's a K param, it's always required, but
r could be replaced by several parameters.  So r should be last
before the N param, so that apply can easily be used to partial
the function, e.g.
   (apply f K params)
to create a single-arg function of N, instead of calling
   (f K params N)